refactor: moved to new gtfs models file
