+++ services/gtfs-processor/src/models/__init__.py
@@ -1,3 +1,4 @@
 from .gtfs_enums import *
+from .gtfs_models import *
 
-__all__ = ["gtfs_enums"]
+__all__ = ["gtfs_enums", "gtfs_models"]
